pyproject.toml
[project]
name = "research-assistant-demo"
version = "0.1.0"
description = "A research assistant demo application powered by Modal, Exa, Logfire"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "chromadb>=1.5.5",
    "exa-py>=2.10.1",
    "google-genai>=1.68.0",
    "logfire[google-genai]>=4.30.0",
    "modal>=1.4.0",
    "mpxpy>=0.0.20",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "pydantic>=2.12.5",
    "requests>=2.33.0",
    "numpy>=2.4.3",
    "pandas>=2.0",
    "jupyter>=1.1.1",
    "prompt-toolkit>=3.0.52",
    "pydantic-ai-slim[google,logfire]>=1.67.0",
    "rich>=14.3.3",
    "textual>=8.1.1",
]

[dependency-groups]
dev = [
    "mypy>=1.10",
    "pydantic-ai>=1.72.0",
    "ruff>=0.15.7",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
]

[tool.uv]
default-groups = []

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["research_assistant"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
